Moving flexibility (concept project) to legacy
