+++ legacy/flexibility_model/electric_and_thermal_grid_planning/concept_parameters.py
