+++ legacy/flexibility_model/electric_and_thermal_grid_planning/get_initial_network.py
