+++ legacy/flexibility_model/electric_and_thermal_grid_planning/plot_functions.py
@@ -1,6 +1,5 @@
 import matplotlib.pyplot as plt
-import get_initial_network as gia
-from cea.optimization.flexibility_model.electric_and_thermal_grid_planning.pyomo_multi_linetype import initial_network
+from legacy.flexibility_model.electric_and_thermal_grid_planning import initial_network
 from cea.technologies.network_layout.substations_location import calc_substation_location
 import re
