+++ legacy/flexibility_model/electric_and_thermal_grid_planning/process_results.py
