+++ legacy/flexibility_model/electric_and_thermal_grid_planning/pyomo_multi_linetype.py
