+++ legacy/flexibility_model/model_building/building_extract_cea_data.py
