+++ legacy/flexibility_model/model_building/building_setup_district.py
