+++ legacy/flexibility_model/model_building/building_utils.py
