+++ legacy/flexibility_model/mpc_building/operation_preprocess_optimization.py
