+++ legacy/flexibility_model/mpc_building/operation_write_results.py
