+++ legacy/flexibility_model/mpc_district/planning_and_operation_write_results.py
